Move ProgressiveFileStream to Controller
